+++ src/Quarter.Core/Repositories/ActivityRepository.cs
@@ -1,16 +1,35 @@
 using System;
 using System.Collections.Generic;
+using System.Threading;
+using System.Threading.Tasks;
 using Npgsql;
 using Quarter.Core.Models;
+using Quarter.Core.Utils;
 
 namespace Quarter.Core.Repositories
 {
     public interface IActivityRepository : IRepository<Activity>
     {
+        Task<Activity> CreateSandboxActivityAsync(IdOf<Project> projectId, CancellationToken ct);
+    }
+
+    public static class ActivityRepository
+    {
+        public static Task<Activity> CreateSandboxActivityAsync(IActivityRepository self, IdOf<Project> projectId, CancellationToken ct)
+        {
+            var activity = new Activity(
+                projectId,
+                "Your first activity",
+                "An activity is used to register time. Each activity has a color to make it easier to associate to.",
+                Color.FromHexString("#4e4694"));
+            return self.CreateAsync(activity, ct);
+        }
     }
 
     public class InMemoryActivityRepository : InMemoryRepositoryBase<Activity>, IActivityRepository
     {
+        public Task<Activity> CreateSandboxActivityAsync(IdOf<Project> projectId, CancellationToken ct)
+            => ActivityRepository.CreateSandboxActivityAsync(this, projectId, ct);
     }
 
     public class PostgresActivityRepository : PostgresRepositoryBase<Activity>, IActivityRepository
@@ -38,5 +57,8 @@ namespace Quarter.Core.Repositories
 
         protected override NpgsqlParameter? WithAccessCondition()
             => new NpgsqlParameter(UserIdColumnName, _userId.Id);
+
+        public Task<Activity> CreateSandboxActivityAsync(IdOf<Project> projectId, CancellationToken ct)
+            => ActivityRepository.CreateSandboxActivityAsync(this, projectId, ct);
     }
 }