+++ CHANGELOG.md
@@ -6,7 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-### Fixes
+### Features
+- Create initial (sandbox) project and activity when creating a new user.
 - Include project and activity usage in project list item.
 
 ## [0.1.3] - 2022-01-24